+++ core/src/main/scala/org/apache/spark/sql/catalyst/expressions/ParamLiteral.scala
@@ -57,7 +57,9 @@ class ParamLiteral(_value: Any, _dataType: DataType, val pos: Int)
 
 //  override def toString: String = s"pl[${super.toString}]"
 
-  override def hashCode(): Int = 31 * (31 * Objects.hashCode(dataType)) + Objects.hashCode(pos)
+  override def hashCode(): Int = {
+    31 * (31 * Objects.hashCode(dataType)) + Objects.hashCode(pos)
+  }
 
   override def equals(obj: Any): Boolean = obj match {
     case a: AnyRef if this eq a => true
